update admin api for role updates
